// src/app/modules/administracion/enums/enums_sin_identificar/regimen.enums.ts
/**
 * Contiene las url generales para conectar con las apis
 */

export enum Enums_Regimen {
    /**
     *  Enums especificos para los grupos 
     */ 
    regimen = 'bill-contribuyente/', 
    todas = 'obtener-todos-contribuyente',
    crear = 'crear',
    actualizar = 'actualizar',
    eliminar = 'eliminar'
}